pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-serve-optimizer-env"
version = "0.1.0"
description = "LLM Deployment Optimizer - OpenEnv Environment"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core==0.2.2",
    "fastapi==0.135.2",
    "uvicorn[standard]==0.42.0",
    "websockets==16.0",
    "openai==2.29.0",
    "pydantic==2.12.5",
    "python-dotenv==1.2.2",
    "requests==2.32.5",
    "psutil==7.2.2"
]

[project.optional-dependencies]
dev = [
    "pytest==9.0.2"
]

[project.scripts]
server = "llm_serve_optimizer_env.server.app:main"

[tool.hatch.build.targets.wheel]
packages = [".", "server", "data"]

[dependency-groups]
dev = [
    "pytest==9.0.2"
]